requirements.txt
ultralytics
fastapi
uvicorn
paddleocr
paddlepaddle
supabase
python-multipart
python-dotenv